fix(security): gate RC_OAUTH_BASE_URL too; move env helpers out of rico
Per review: also lock the OAuth base URL to production in release builds so
auth can't be redirected, and move envOrDefault/devEnvOrDefault to env.go since
they're shared, not rico-specific.
